CInput stuff

+++ src/XDK/LIBCMT/types_compat.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #ifndef _SIZE_T_DEFINED
-typedef unsigned long size_t; // or whatever PPC XDK used
+typedef unsigned int size_t; // or whatever PPC XDK used
 #define _SIZE_T_DEFINED
 #endif
